backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-jose
passlib[bcrypt]
vosk
python-multipart
python-jose[cryptography] 
pydantic[email]==2.12.5
email-validator==2.3.0
websockets==12.0
vosk==0.3.45
numpy==1.24.3